Three Views Shown
Continuing my exploration with this project.

The current version requires you to click a button to start the
animation. It saves the final render state to local storage as a
compressed json string.

The restore operation for this is not quite working well yet. I am
thinking it might be simpler to do the full serialize deserialize and
storage in Rust rather than trying to pass the objects back from js to
rust (there will be the issue of compressing the string to solve there
though, may need a different library)

Three views are being rendered. The overall size and number of points
has been made smaller but then is blown up in the bottom render by using
lines rather than points